Better defaults for user and group ownership
For chown() and chgrp() operations we were assuming 'nginx' and
'nginx' unless a different value was specified.  While this works
for many cases, it's still annoying when running those as a
non-root user.

So, from now on, we assume the user and group based on the effective
ID.  In other words, the user and group of the current process.