Check for POSIX extension
POSIX extension is not always installed.  If it's not there, some
of the filesystem tasks can throw fatal errors about posix
functions being unknown.  The user needs either either provide all
necessary paramenters (to avoid guessing for current user and group)
or install the php-posix extension.

This commit checks whether or not the posix-extension is installed
before calling the posix_* functions.  This is a better solution than
doing this check in composer, as there is plenty of phake-builder
functionality that can be useful without the filesystem bits.